Refine visualization environments and use base class in elbow
